HUD: Added function for shifting a position
It seems useful to have this, because instead of assigning a fixed
position to GUI elements (assuming that the others are in their own
fixed positions) you have to do a lot of calculations - either hidden in
your personal notes, or in the code (but this can become confusing).

Shifting the position if an element is there or not makes it more
dynamic. Ideally, there would be a "docking" mechanism for separate
elements.