HUD: Display info text
Notes: The adding of getters is not going to be a problem once we have
multiple GUI elements that want to define the same getters (which is
bound to happen :D). Will have to look for a better solution

Further notes:
Fire modes and fire techniques.

In the library there are only fire modes. In CMC there is a fire mode
(which is the same as ammo type), and the fire technique (that
determines HOW the weapon is fired, e.g. burst mode, or if fire mode =
grenades, then fire technique may be smoke grenade or shrapnel, etc.)

Any ideas how we should handle this? Maybe a simple expansion of the
proplist is sufficient (i.e. GetFiremode()->GetMode(),
GetFiremode()->GetTechnique())