Class System: Added a basic security check for properties
Creating setters makes class creation safe against renaming the internal
properties, but it also requires you to call these functions somehow.
Instead, it seemed more fruitful to assert that the class does not
implement properties that are not defined by the parent.

This seems to be useful in OC in general, so I already placed the check
in the legacy folder.