GUI Elements: Implement a list menu
Shamelessly copied from the OC original MenuStyle_List by Zapper. Might
be useful to use objects instead of proplists after all (but only for
the prototype), because function overloading seems to work differently
for objects and proplists (i.e. the proplists have not inherited
function?). Still, I would not create an actual object in that case,
just use it as a prototype.