Respawn: Create class selection tabs
This is unfinished. The final result should have callbacks for filling
the tabs, flexible entry assignment as in other menus.

This change breaks the game, because there will be no way to enter the
map at the moment (not that it makes much of a difference in the current
state :D)