Selection times: Basic hammerspace animation
This is far from perfect, but it is the basic mechanic as long as there
are no real new animations. Basic selection time is 35 frames now.

It was disappointing that there seems to be no other way than setting
the transformation manually. Maybe setting the animation with linear
scaling and setting the animation position would help.

Real values for upper and lower hand bone will be there soon. What still
bothers me, however, is that the displayed item will be changed
immediately. This might solve itself soon, because I want to override
HasHandAction and certain controls to be blocked as long as the
selection time is active.