Reload: Pistol, SMG
Sound intervals are not that great, it seems? Found a bug in the shooter
library. Apparently now the cooldown is applied after every full-auto
shot (most noticeable in the machine gun), but it should be applied only
once you let go of the fire button, or at the end of a burst.