Library update
Created a branch for shooter (or other) library updates. I will make
changes here as I simultaneously update the library in its own
repository. Then, once everything is in the desired state I will merge
it back to the main branch.

Will not update the library file on this branch though, because I
include the decompiled library locally.