BoobyTrap: Defuse via collection
Has to be tested during gameplay. In CR you had to grab the trap and
press up, which takes some time and feels like effort / risk to the
player.
At the moment you can defuse the trap by simply collecting it, but to do
so you have to press a button now, so might be good, might be bad.