Issue #11: Ported blood system
This is a generic issue with the damage callbacks. Sure, we can add a
Hazard-like custom damage system and callbacks again, as in the old
project, but that has always the problem that other packs are
incompatible (not so much of a problem in general, but actually changing
the engine, because there is a need for the change and the current state
causing problems for years makes me think that we should change the
engine instead).